Test moved from separate folder to the Provisioner test
