--- tests/hibernation/test/hibernation/main_test.go
@@ -1,44 +1,0 @@
-package provisioner
-
-import (
-	"os"
-	"testing"
-
-	"github.com/sirupsen/logrus"
-
-	"github.com/kyma-project/control-plane/tests/hibernation/test/testkit"
-)
-
-var (
-	config    testkit.TestConfig
-	testSuite *TestSuite
-)
-
-func TestMain(m *testing.M) {
-	var err error
-
-	config, err = testkit.ReadConfig()
-	if err != nil {
-		logrus.Errorf("Failed to read configuration: %s", err.Error())
-		os.Exit(1)
-	}
-
-	code := runTests(m)
-	os.Exit(code)
-}
-
-func runTests(m *testing.M) int {
-	var err error
-	testSuite, err = NewTestSuite(config)
-	if err != nil {
-		logrus.Errorf("Failed to create new test suite: %s", err.Error())
-		return 1
-	}
-	defer testSuite.Cleanup()
-
-	if err = testSuite.Setup(); err != nil {
-		logrus.Errorf("Failed to setup tests environment: %s", err.Error())
-		return 1
-	}
-	return m.Run()
-}
